fix: Correct inline asm for WriteNvgChannelIdx
Correct issue with register being used from a 32 bit value.

Bug 3610319

Reviewed-on: https://git-master.nvidia.com/r/c/tegra/bootloader/uefi/edk2-nvidia/+/2699777
Tested-by: Jeff Brasen <jbrasen@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ Silicon/NVIDIA/Library/NvgLib/NvgLib.c
@@ -19,7 +19,7 @@
   asm volatile ("mrs %0, "#reg : "=r"(var) : : "memory", "cc"); \
 } while (FALSE)
 
-static inline void WriteNvgChannelIdx(UINT32 Channel)
+static inline void WriteNvgChannelIdx(UINT64 Channel)
 {
   asm volatile ("msr s3_0_c15_c1_2, %0" : : "r"(Channel) : "memory", "cc");
 }
